Move Result class into independent file
We prefer more but shortish and specialized files.
